Fix/supress cppcheck warnings
- 32 confirmed false positives — all traced to cppcheck not tracking
  ASSERT(...) preconditions (which compile to nothing in release
  builds) across accessor calls, constructor→onExecute time separation,
  or macro bodies.
- src/app/file/file.cpp: fixed a signed/unsigned comparison with an
  explicit static_cast<int>.
- src/app/document_api.cpp: added ASSERT(sprite) to adjustFrameTags(),
  matching the precondition style used by every other DocumentApi
  method in the file.
- src/doc/palette.cpp: clamped ncolors to std::max(ncolors, 0) before
  resize() in both the constructor and Palette::resize(), so a
  malformed .ico/.ase file with a negative color count degrades to an
  empty palette in release builds instead of converting to a huge
  size_t and throwing/crashing.
  The ASSERT still catches it immediately in debug builds.

+++ src/app/file/file.cpp
@@ -357,7 +357,7 @@ FileOp* FileOp::createSaveDocumentOperation(const Context* context,
       for (int c = 0; c < pal->size(); ++c)
       {
         int a = rgba_geta(pal->getEntry(c));
-        if (oneAlpha && c == transparentColor)
+        if (oneAlpha && c == static_cast<int>(transparentColor))
         {
           continue;
         }
